nixbot: rename modules to disambiguate scheduling and rerun concerns
The package had three modules sharing the "schedul*" prefix that cover
unrelated domains, plus two "re*" modules whose docstrings overlapped,
making it unclear where logic lived:

  scheduler.py      -> build_scheduler.py  (build-attribute scheduling)
  scheduled.py      -> schedules.py        (onSchedule model + store)
  scheduled_runs.py -> schedule_runner.py  (due-effect execution loop)
  reruns.py         -> rerun_exec.py       (orchestrator rerun primitives)
  restarts.py       -> restart_dispatch.py (work-queue restart entry points)

Pure rename, no behavior change. The build-attribute scheduler is
unrelated to the onSchedule (Hercules effect) machinery; the new names
make that boundary explicit and stop the scheduler/scheduled collision.
