+++ nixbot/nixbot/tests/test_pipeline.py
@@ -9,6 +9,12 @@ from typing import TYPE_CHECKING
 
 import pytest
 
+from nixbot.build_scheduler import (
+    AttributeStatus,
+    BuildOutcome,
+    JobScheduler,
+    ScheduleResult,
+)
 from nixbot.executor import (
     BuildSettings,
     FairScheduler,
@@ -18,12 +24,6 @@ from nixbot.executor import (
 )
 from nixbot.nix_eval import EvalRunner, EvalSettings
 from nixbot.repo_config import BranchConfig
-from nixbot.scheduler import (
-    AttributeStatus,
-    BuildOutcome,
-    JobScheduler,
-    ScheduleResult,
-)
 
 if TYPE_CHECKING:
     from pathlib import Path
